fix: split bfb_rshim_copier into its own module
