+++ crates/preingestion-manager/src/bfb_rshim_copier.rs
@@ -1,0 +1,255 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use std::sync::Arc;
+use std::time::Duration;
+
+use forge_secrets::credentials::{CredentialKey, CredentialReader, Credentials};
+use tokio::fs::{self, File};
+use tokio::io::{AsyncReadExt, AsyncWriteExt};
+use tokio::sync::Mutex;
+
+const UNIFIED_PREINGESTION_BFB_PATH: &str =
+    "/forge-boot-artifacts/blobs/internal/aarch64/preingestion_unified_update.bfb";
+const PREINGESTION_BFB_PATH: &str = "/forge-boot-artifacts/blobs/internal/aarch64/preingestion.bfb";
+
+#[derive(Debug, thiserror::Error)]
+pub(crate) enum BfbRshimCopyError {
+    #[error("Missing credential {key}: {cause}")]
+    MissingCredentials { key: String, cause: String },
+    #[error("Secrets engine error occurred: {cause}")]
+    SecretsEngineError { cause: String },
+    #[error("Error: {details}")]
+    Other { details: String },
+}
+
+pub(crate) struct BfbRshimCopier {
+    credential_reader: Option<Arc<dyn CredentialReader>>,
+    bfb_file_lock: Arc<Mutex<()>>,
+}
+
+impl BfbRshimCopier {
+    pub(crate) fn new(credential_reader: Option<Arc<dyn CredentialReader>>) -> Self {
+        Self {
+            credential_reader,
+            bfb_file_lock: Arc::new(Mutex::new(())),
+        }
+    }
+
+    fn valid_bmc_password(credentials: &Credentials) -> bool {
+        let (_, password) = match credentials {
+            Credentials::UsernamePassword { username, password } => (username, password),
+        };
+
+        !password.is_empty()
+    }
+
+    async fn get_bmc_root_credentials(
+        &self,
+        credential_key: &CredentialKey,
+    ) -> Result<Credentials, BfbRshimCopyError> {
+        let Some(credential_reader) = &self.credential_reader else {
+            return Err(BfbRshimCopyError::MissingCredentials {
+                key: credential_key.to_key_str().to_string(),
+                cause: "credential reader is not configured".to_string(),
+            });
+        };
+
+        match credential_reader.get_credentials(credential_key).await {
+            Ok(Some(credentials)) => {
+                if !Self::valid_bmc_password(&credentials) {
+                    return Err(BfbRshimCopyError::Other {
+                        details: format!(
+                            "vault does not have a valid password entry at {}",
+                            credential_key.to_key_str()
+                        ),
+                    });
+                }
+
+                Ok(credentials)
+            }
+            Ok(None) => Err(BfbRshimCopyError::MissingCredentials {
+                key: credential_key.to_key_str().to_string(),
+                cause: "No credentials exists".to_string(),
+            }),
+            Err(err) => Err(BfbRshimCopyError::SecretsEngineError {
+                cause: err.to_string(),
+            }),
+        }
+    }
+
+    async fn is_rshim_enabled(
+        &self,
+        bmc_ip_address: std::net::SocketAddr,
+        credentials: Credentials,
+    ) -> Result<bool, BfbRshimCopyError> {
+        let (username, password) = match credentials {
+            Credentials::UsernamePassword { username, password } => (username, password),
+        };
+
+        forge_ssh::ssh::is_rshim_enabled(bmc_ip_address, username, password)
+            .await
+            .map_err(|err| BfbRshimCopyError::Other {
+                details: format!("failed query RSHIM status on on {bmc_ip_address}: {err}"),
+            })
+    }
+
+    async fn enable_rshim(
+        &self,
+        bmc_ip_address: std::net::SocketAddr,
+        credentials: Credentials,
+    ) -> Result<(), BfbRshimCopyError> {
+        let (username, password) = match credentials {
+            Credentials::UsernamePassword { username, password } => (username, password),
+        };
+
+        forge_ssh::ssh::enable_rshim(bmc_ip_address, username, password)
+            .await
+            .map_err(|err| BfbRshimCopyError::Other {
+                details: format!("failed enable RSHIM on {bmc_ip_address}: {err}"),
+            })
+    }
+
+    async fn check_and_enable_rshim(
+        &self,
+        bmc_ip_address: std::net::SocketAddr,
+        credentials: &Credentials,
+    ) -> Result<(), BfbRshimCopyError> {
+        let mut i = 0;
+        while i < 3 {
+            if !self
+                .is_rshim_enabled(bmc_ip_address, credentials.clone())
+                .await?
+            {
+                tracing::warn!("RSHIM is not enabled on {bmc_ip_address}");
+                self.enable_rshim(bmc_ip_address, credentials.clone())
+                    .await?;
+
+                // Sleep for 10 seconds before checking again
+                tokio::time::sleep(Duration::from_secs(10)).await;
+                i += 1;
+            } else {
+                return Ok(());
+            }
+        }
+
+        Err(BfbRshimCopyError::Other {
+            details: format!("could not enable RSHIM on {bmc_ip_address}"),
+        })
+    }
+
+    async fn create_unified_preingestion_bfb(
+        &self,
+        username: &str,
+        password: &str,
+    ) -> Result<(), BfbRshimCopyError> {
+        let _lock = self.bfb_file_lock.lock().await;
+
+        if fs::metadata(UNIFIED_PREINGESTION_BFB_PATH).await.is_err() {
+            tracing::info!("Writing {UNIFIED_PREINGESTION_BFB_PATH}");
+            let bf_cfg_contents = format!(
+                "BMC_USER=\"{username}\"\nBMC_PASSWORD=\"{password}\"\nBMC_REBOOT=\"yes\"\nCEC_REBOOT=\"yes\"\n"
+            );
+
+            let mut preingestion_bfb = File::open(PREINGESTION_BFB_PATH).await.map_err(|err| {
+                BfbRshimCopyError::Other {
+                    details: format!("failed to open {PREINGESTION_BFB_PATH}: {err}"),
+                }
+            })?;
+
+            let mut unified_bfb =
+                File::create(UNIFIED_PREINGESTION_BFB_PATH)
+                    .await
+                    .map_err(|err| BfbRshimCopyError::Other {
+                        details: format!("failed to create {UNIFIED_PREINGESTION_BFB_PATH}: {err}"),
+                    })?;
+
+            let mut buffer = vec![0; 1024 * 1024].into_boxed_slice(); // 1 MB buffer
+
+            tracing::info!("Writing BFB to {UNIFIED_PREINGESTION_BFB_PATH}");
+            loop {
+                let n = preingestion_bfb.read(&mut buffer).await.map_err(|err| {
+                    BfbRshimCopyError::Other {
+                        details: format!("failed to read BFB: {err}"),
+                    }
+                })?;
+
+                if n == 0 {
+                    break;
+                }
+
+                unified_bfb.write_all(&buffer[..n]).await.map_err(|err| {
+                    BfbRshimCopyError::Other {
+                        details: format!(
+                            "failed to write BFB to {UNIFIED_PREINGESTION_BFB_PATH}: {err}"
+                        ),
+                    }
+                })?;
+            }
+
+            tracing::info!("Writing bf.cfg to {UNIFIED_PREINGESTION_BFB_PATH}");
+
+            unified_bfb
+                .write_all(bf_cfg_contents.as_bytes())
+                .await
+                .map_err(|err| BfbRshimCopyError::Other {
+                    details: format!("failed to write bf.cfg: {err}"),
+                })?;
+
+            unified_bfb
+                .sync_all()
+                .await
+                .map_err(|err| BfbRshimCopyError::Other {
+                    details: format!("failed to flush {UNIFIED_PREINGESTION_BFB_PATH}: {err}"),
+                })?;
+        }
+
+        Ok(())
+    }
+
+    pub(crate) async fn copy_bfb_to_dpu_rshim(
+        &self,
+        bmc_ip_address: std::net::SocketAddr,
+        credential_key: &CredentialKey,
+        is_bf2: bool,
+    ) -> Result<(), BfbRshimCopyError> {
+        let credentials = self.get_bmc_root_credentials(credential_key).await?;
+        let (username, password) = match credentials.clone() {
+            Credentials::UsernamePassword { username, password } => (username, password),
+        };
+
+        self.create_unified_preingestion_bfb(&username, &password)
+            .await?;
+
+        self.check_and_enable_rshim(bmc_ip_address, &credentials)
+            .await?;
+
+        forge_ssh::ssh::copy_bfb_to_bmc_rshim(
+            bmc_ip_address,
+            username,
+            password,
+            UNIFIED_PREINGESTION_BFB_PATH.to_string(),
+            is_bf2,
+        )
+        .await
+        .map_err(|err| BfbRshimCopyError::Other {
+            details: format!(
+                "failed to copy BFB from {UNIFIED_PREINGESTION_BFB_PATH} to BMC RSHIM on {bmc_ip_address}: {err}"
+            ),
+        })
+    }
+}
